Day 8
A bit slower than I'd like, but still well under a half a second and I built the algorithm myself rather than looking up an implementation of Disjoint Set Union or Kruskal's or whatever, so I'm just going to roll with it. Doing so much array scanning, clipping, and appending is likely the source of the lower-than-it-could-be performance, but I feel it's readable now and this input doesn't require further optimizations.